working state for 1023 chars but with 1024 bufferoverflow

+++ new_minitalk/server.c
@@ -6,7 +6,7 @@
 /*   By: pharbst <pharbst@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 14:45:06 by pharbst           #+#    #+#             */
-/*   Updated: 2022/11/23 04:13:21 by pharbst          ###   ########.fr       */
+/*   Updated: 2022/11/23 16:57:39 by pharbst          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,13 +30,15 @@ void	sig_handler(int sig, siginfo_t *siginfo, void *context)
 	static char	*c[4194304][(INT_MAX / 1024) / 1024];
 
 	pid = siginfo->si_pid;
+	if (pid < 1)
+		return ;
 	(void)context;
 	if (flag[pid] == 0 && sig == SIGUSR1)
 	{
 		ft_printf("Connection established with %d\n", pid);
 		flag[pid] = 1;
 		kill(pid, SIGUSR1);
-		usleep(100);
+		// usleep(100);
 	}
 	else if (flag[pid] == 1)
 	{
@@ -47,14 +49,24 @@ void	sig_handler(int sig, siginfo_t *siginfo, void *context)
 		if (!c[pid][(i[pid] / 8) / 1024])
 			c[pid][(i[pid] / 8) / 1024] = ft_calloc(1024, sizeof(char));
 		if (sig == SIGUSR1)
+		{
 			c[pid][(i[pid] / 8) / 1024][(i[pid] / 8) % 1024] = (c[pid][(i[pid] / 8) / 1024][(i[pid] / 8) % 1024] << 1) + 1;
+			// ft_printf("1");
+		}
 		if (sig == SIGUSR2)
+		{
 			c[pid][(i[pid] / 8) / 1024][(i[pid] / 8) % 1024] = (c[pid][(i[pid] / 8) / 1024][(i[pid] / 8) % 1024] << 1);
+			// ft_printf("0");
+		}
 		j = 0;
-		if (i[pid] / 8 != 0 && i[pid] % 8 == 0 && c[pid][(i[pid] / 8) / 1024][(i[pid] / 8) % 1024] == '\0')
+		// ft_printf("\ni[pid] / 8 = %d && i[pid] %% 8 = %d && c[pid][(i[pid] / 8) / 1024][(i[pid] / 8) %% 1024] = %c\n", i[pid] / 8, i[pid] % 8, c[pid][(i[pid] / 8) / 1024][(i[pid] / 8) % 1024]);
+		if (i[pid] / 8 != 0 && i[pid] % 8 == 7 && c[pid][(i[pid] / 8) / 1024][(i[pid] / 8) % 1024] == '\0')
 		{
 			while (c[pid][j])
+			{
 				print_block(c[pid][j++]);
+				// ft_printf("block printed\n");
+			}
 			write(1, "\n", 1);
 			j = 0;
 			while (c[pid][j])
@@ -69,12 +81,10 @@ void	sig_handler(int sig, siginfo_t *siginfo, void *context)
 int	main()
 {
 	struct sigaction	sa;
-	{
-		sa.sa_flags = SA_SIGINFO;
-		sa.sa_sigaction = sig_handler;
-		sigaction(SIGUSR1, &sa, NULL);
-		sigaction(SIGUSR2, &sa, NULL);
-	};
+	sa.sa_flags = SA_SIGINFO;
+	sa.sa_sigaction = sig_handler;
+	sigaction(SIGUSR1, &sa, NULL);
+	sigaction(SIGUSR2, &sa, NULL);
 	printf("Server is running\n");
 	fflush(stdout);
 	ft_printf("Server PID: %d\n", getpid());
